% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nunnaly_cpp.R
\name{def_GH}
\alias{def_GH}
\title{Ben Mayo code}
\usage{
def_GH(
  G = 0:25,
  Hp = 0:12,
  D0 = c(0, 1, 4),
  D1 = c(0, 2, 7),
  Len = length(G),
  Hlen = length(Hp),
  Dlen = length(D),
  F_mat = matrix(exp((0+2i) * pi * G \%*\% t(G)/Len), ncol = Len),
  D = c(2 * D0, (1 + 2 * D1)),
  Phi = F_mat[D + 1, ]/sqrt(Dlen),
  Psi = matrix(rep(0, (Len * Dlen)), ncol = Len),
  TwoNorm = matrix(rep(0, Hlen^2), ncol = Hlen),
  FroNorm = TwoNorm
)
}
\description{
Ben Mayo code
}
