% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv.R
\name{pv}
\alias{pv}
\title{Compute present value of money}
\usage{
pv(FV, r, n = 5)
}
\arguments{
\item{FV}{A \code{numeric} vector forget what this stands for}

\item{r}{A \code{numeric} value for the tate of return}

\item{n}{An \code{integer} for the number of years}
}
\value{
A numeric vector of the same length as FV
}
\description{
Compute present value of money
}
\examples{
\dontrun{
pv(FV = c(0.3,0.5), r = 0.25)
}
}
