% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_py.R
\name{test_py}
\alias{test_py}
\title{Test out running Python}
\usage{
test_py(sampsize, nsim, exp = T, param = 0.3)
}
\arguments{
\item{sampsize}{The size of the sample}

\item{nsim}{The number of simulations}
}
\value{
A seaborn kdeplot (kernel density estimate plot)
}
\description{
Test out running Python
}
