% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_sampler.R
\name{test_sampler}
\alias{test_sampler}
\title{Test Python function `sampler`}
\usage{
test_sampler(df, colname, N, alpha, mu_0)
}
\arguments{
\item{df}{A \code{data.frame}-class object}

\item{colname}{Character string denoting the name of the column within \code{df} to work with}

\item{N}{Integer for the size of the sample}

\item{alpha}{Numeric value for the significance}

\item{mu_0}{The value of the mean for the Null hypothesis}
}
\value{
A named list
}
\description{
Test Python function `sampler`
}
